feat: oauth sec implementation for all modules and request/response format problems fixed

+++ backend/app/modules/inventory/router.py
@@ -1,10 +1,13 @@
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, status, Query
 from sqlmodel import select
 from uuid import UUID
 from app.modules.inventory.service import ProductService
 from app.modules.inventory.schemas import ProductCreate, ProductRead, ProductUpdate
 from app.modules.inventory.models import Product
 from app.core.db import SessionDep
+from app.shared.dependencies import get_current_user
+from app.modules.auth.schemas import TokenData
+from app.shared.enums import UserRoleEnum
 router = APIRouter()
 
 def get_service(session: SessionDep) -> ProductService:
@@ -19,7 +22,13 @@ def get_service(session: SessionDep) -> ProductService:
 async def create_product(
     product_input: ProductCreate, 
     service: ProductService = Depends(get_service),
+    token_data: TokenData = Depends(get_current_user)
     ):
+    if token_data.role != UserRoleEnum.ADMIN:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="No tienes permisos para crear productos"
+        )
     return await service.create_product_as_service(product_input=product_input)
 
 @router.get(
@@ -31,8 +40,11 @@ async def create_product(
 async def get_product(
     service: ProductService = Depends(get_service),
     search: str | None = None, 
+    offset: int = 0,
+    limit: int = Query(default=20, ge=1, le=100),
+    token_data: TokenData = Depends(get_current_user)
 ):
-    return await service.get_product_as_service(search=search)
+    return await service.get_product_as_service(search=search, offset=offset, limit=limit)
 
 @router.patch(
     "/{product_id}",
@@ -44,7 +56,13 @@ async def update_product(
     product_id: UUID, 
     product_input: ProductUpdate, 
     service: ProductService = Depends(get_service),
+    token_data: TokenData = Depends(get_current_user)
 ):
+    if token_data.role != UserRoleEnum.ADMIN:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="No tienes permisos para actualizar productos"
+        )
     return await service.update_product_as_service(product_id=product_id, product_input=product_input)
 
 @router.delete(
@@ -55,5 +73,11 @@ async def update_product(
 async def delete_product(
     product_id: UUID,
     service: ProductService = Depends(get_service),
+    token_data: TokenData = Depends(get_current_user)
 ):
+    if token_data.role != UserRoleEnum.ADMIN:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="No tienes permisos para eliminar productos"
+        )
     return await service.delete_product_as_service(product_id=product_id)
